cleanup: remove purchase requirement related code
