--- src/components/[guild]/Requirements/components/GuildCheckout/components/PaymentCurrencyPicker/index.ts
@@ -1,3 +1,0 @@
-import PaymentCurrencyPicker from "./PaymentCurrencyPicker"
-
-export default PaymentCurrencyPicker
