--- src/components/[guild]/Requirements/components/GuildCheckout/components/PaymentMethodButtons.tsx
@@ -1,51 +1,0 @@
-import { usePostHogContext } from "@/components/Providers/PostHogProvider"
-import { Box, ButtonGroup, Tooltip } from "@chakra-ui/react"
-import useGuild from "components/[guild]/hooks/useGuild"
-import Button from "components/common/Button"
-
-const PaymentMethodButtons = () => {
-  const { captureEvent } = usePostHogContext()
-  const { urlName } = useGuild()
-
-  const onPayWithCryptoClick = () =>
-    captureEvent("Click: Pay with crypto (GuildCheckout)", {
-      guild: urlName,
-    })
-
-  const onPayWithCardClick = () =>
-    captureEvent("Click: Pay with card (GuildCheckout)", {
-      guild: urlName,
-    })
-
-  return (
-    <ButtonGroup size="sm" w="full" mb="8">
-      <Button
-        autoFocus={false}
-        colorScheme="blue"
-        variant="subtle"
-        w="full"
-        borderRadius="md"
-        onClick={onPayWithCryptoClick}
-      >
-        Pay with crypto
-      </Button>
-
-      <Tooltip label="Coming soon" placement="top" hasArrow>
-        <Box w="full" onClick={onPayWithCardClick}>
-          <Button
-            autoFocus={false}
-            variant="subtle"
-            w="full"
-            borderRadius="md"
-            isDisabled
-            _hover={""}
-          >
-            Pay with card
-          </Button>
-        </Box>
-      </Tooltip>
-    </ButtonGroup>
-  )
-}
-
-export default PaymentMethodButtons
