--- src/components/[guild]/Requirements/components/GuildCheckout/components/buttons/PurchaseAllowanceButton.tsx
@@ -1,109 +1,0 @@
-import { usePostHogContext } from "@/components/Providers/PostHogProvider"
-import { Collapse, Icon, Tooltip } from "@chakra-ui/react"
-import { Check, Question, Warning } from "@phosphor-icons/react"
-import useAllowance from "components/[guild]/Requirements/components/GuildCheckout/hooks/useAllowance"
-import useGuild from "components/[guild]/hooks/useGuild"
-import Button from "components/common/Button"
-import useTokenData from "hooks/useTokenData"
-import { NULL_ADDRESS, TOKEN_BUYER_CONTRACTS } from "utils/guildCheckout/constants"
-import { useChainId } from "wagmi"
-import { CHAIN_CONFIG, Chain, Chains } from "wagmiConfig/chains"
-import { useRequirementContext } from "../../../RequirementContext"
-import usePrice from "../../hooks/usePrice"
-import { useGuildCheckoutContext } from "../GuildCheckoutContext"
-
-const PurchaseAllowanceButton = (): JSX.Element => {
-  const { captureEvent } = usePostHogContext()
-  const { urlName } = useGuild()
-
-  const requirement = useRequirementContext()
-  // TODO: remove typecast once we implement schemas for ERC20 requirements
-  const requirementChain = requirement.chain as Chain
-  const requirementChainId = Chains[requirementChain]
-  const { pickedCurrency } = useGuildCheckoutContext()
-
-  const chainId = useChainId()
-
-  const {
-    data: { symbol, name },
-  } = useTokenData(requirementChain, pickedCurrency)
-  const nativeCurrency = CHAIN_CONFIG[Chains[chainId]]?.nativeCurrency
-  const isNativeCurrencyPicked = pickedCurrency === NULL_ADDRESS
-
-  const tokenSymbol = isNativeCurrencyPicked ? nativeCurrency.symbol : symbol
-  const tokenName = isNativeCurrencyPicked ? nativeCurrency.name : name
-
-  const {
-    data: { maxPriceInWei },
-    isValidating: isPriceLoading,
-  } = usePrice()
-
-  const {
-    allowance,
-    isAllowanceLoading,
-    isAllowing,
-    allowanceError,
-    allowSpendingTokens,
-  } = useAllowance(pickedCurrency, TOKEN_BUYER_CONTRACTS[Chains[chainId]]?.address)
-
-  const isEnoughAllowance =
-    typeof maxPriceInWei === "bigint" && typeof allowance === "bigint"
-      ? maxPriceInWei <= allowance
-      : false
-
-  const onClick = () => {
-    allowSpendingTokens()
-    captureEvent("Click: PurchaseAllowanceButton (GuildCheckout)", {
-      guild: urlName,
-    })
-  }
-
-  return (
-    <Collapse
-      in={
-        pickedCurrency &&
-        chainId === requirementChainId &&
-        !isNativeCurrencyPicked &&
-        !isEnoughAllowance
-      }
-    >
-      <Button
-        size="lg"
-        colorScheme={allowanceError ? "red" : "blue"}
-        isDisabled={isEnoughAllowance}
-        isLoading={isPriceLoading || isAllowanceLoading || isAllowing}
-        loadingText={
-          isPriceLoading || isAllowanceLoading
-            ? "Checking allowance"
-            : isAllowing
-              ? "Allowing"
-              : "Check your wallet"
-        }
-        onClick={onClick}
-        w="full"
-        leftIcon={
-          allowanceError ? (
-            <Icon as={Warning} />
-          ) : isEnoughAllowance ? (
-            <Icon as={Check} />
-          ) : null
-        }
-        rightIcon={
-          !isEnoughAllowance && (
-            <Tooltip
-              label={`You have to give the Guild smart contracts permission to use your ${tokenName}. You only have to do this once per token.`}
-            >
-              <Icon as={Question} />
-            </Tooltip>
-          )
-        }
-      >
-        {allowanceError
-          ? "Couldn't fetch allowance"
-          : `Allow Guild to use your ${tokenSymbol}`}
-      </Button>
-    </Collapse>
-  )
-}
-
-export default PurchaseAllowanceButton
